LazyLearning_SVM/requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.1
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.7
mlxtend==0.23.4
numpy==2.3.3
packaging==25.0
pandas==2.3.3
patsy==1.0.2
pillow==11.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.2
scikit-posthocs==0.11.4
scipy==1.16.2
seaborn==0.13.2
six==1.17.0
sklearn-relief==1.0.0b2
statsmodels==0.14.5
threadpoolctl==3.6.0
tqdm==4.67.1
tzdata==2025.2
